+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.0.1.0
+
+Simple strings (scalars) are now output unquoted.
+
+Added:
+  - encodeQuoted: Encodes with all keys/strings quoted
+  - encodeQuotedDocuments: Encodes documents with all keys/strings quoted
+
 # 1.0.0.0
 
 Initial version